move KspExtensions package
